update utility function
